* ModuleManager first version * Uploader incomplete
