+++ Client/Program.cs
@@ -8,4 +8,6 @@ builder.RootComponents.Add<HeadOutlet>("head::after");
 
 var httpClient = new HttpClient { BaseAddress = new Uri(builder.HostEnvironment.BaseAddress) };
 builder.Services.AddScoped(sp => httpClient);
+builder.Services.AddScoped<IComponentRepository, ComponentRepository>();
+
 await builder.Build().RunAsync();
