+++ Client/Shared/NavMenu.razor
@@ -15,8 +15,8 @@
             </NavLink>
         </div>
         <div class="nav-item px-3">
-            <NavLink class="nav-link" href="counter">
-                <span class="oi oi-plus" aria-hidden="true"></span> Counter
+            <NavLink class="nav-link" href="module">
+                <span class="oi oi-plus" aria-hidden="true"></span> Module Manager
             </NavLink>
         </div>
         <div class="nav-item px-3">
